+++ src/test/java/org/scion/jpan/api/ScmpSenderTest.java
@@ -22,10 +22,7 @@ import java.nio.ByteBuffer;
 import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Predicate;
 import java.util.function.Supplier;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
@@ -354,94 +351,6 @@ public class ScmpSenderTest {
       List<Path> paths = service.getPaths(dstIA, new InetSocketAddress(addr, port));
       return paths.get(0);
     } catch (UnknownHostException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @Test
-  void setUpScmpResponder_echo() throws IOException, InterruptedException {
-    MockNetwork.startTiny();
-    MockScmpHandler.stop(); // Shut down SCMP handler
-    Path path = getPathTo112(InetAddress.getLoopbackAddress());
-    // sender is in 110; responder is in 112
-    Thread responder = null;
-    try (ScmpSender sender = Scmp.newSenderBuilder().build()) {
-      sender.setScmpErrorListener(scmpMessage -> errors.add(scmpMessage.getTypeCode().getText()));
-      sender.setOption(ScionSocketOptions.SCION_API_THROW_PARSER_FAILURE, true);
-
-      // start responder
-      responder = startResponderThread(null);
-
-      // send request
-      for (int i = 0; i < 10; i++) {
-        Scmp.EchoMessage msg = sender.sendEchoRequest(path, ByteBuffer.allocate(0));
-        assertNotNull(msg);
-        assertFalse(msg.isTimedOut(), "i=" + i);
-        assertEquals(Scmp.TypeCode.TYPE_129, msg.getTypeCode());
-      }
-    } finally {
-      stopResponderThread(responder);
-      MockNetwork.stopTiny();
-    }
-  }
-
-  @Test
-  void setUpScmpResponder_echo_blocked() throws IOException, InterruptedException {
-    MockNetwork.startTiny();
-    MockScmpHandler.stop(); // Shut down SCMP handler
-    Path path = getPathTo112(InetAddress.getLoopbackAddress());
-    // sender is in 110; responder is in 112
-    Thread responder = null;
-    try (ScmpSender sender = Scmp.newSenderBuilder().build()) {
-      sender.setScmpErrorListener(scmpMessage -> errors.add(scmpMessage.getTypeCode().getText()));
-      sender.setOption(ScionSocketOptions.SCION_API_THROW_PARSER_FAILURE, true);
-
-      // start responder
-      AtomicInteger dropCount = new AtomicInteger();
-      responder = startResponderThread((echoMsg) -> dropCount.incrementAndGet() < -42);
-
-      // send request
-      sender.setTimeOut(100);
-      for (int i = 0; i < 2; i++) {
-        Scmp.EchoMessage result = sender.sendEchoRequest(path, ByteBuffer.allocate(0));
-        assertTrue(result.isTimedOut());
-        assertEquals(100 * 1_000_000, result.getNanoSeconds());
-        assertEquals(i, result.getSequenceNumber());
-      }
-      assertEquals(2, dropCount.get());
-    } finally {
-      stopResponderThread(responder);
-      MockNetwork.stopTiny();
-    }
-  }
-
-  private Thread startResponderThread(Predicate<Scmp.EchoMessage> predicate)
-      throws InterruptedException {
-    CountDownLatch barrier = new CountDownLatch(1);
-    Thread t = new Thread(() -> scmpResponder(barrier, predicate));
-    t.start();
-    barrier.await();
-    Thread.sleep(50);
-    return t;
-  }
-
-  private void stopResponderThread(Thread responder) throws InterruptedException {
-    if (responder != null) {
-      responder.join(100);
-      responder.interrupt(); // just in case.
-    }
-  }
-
-  private void scmpResponder(CountDownLatch barrier, Predicate<Scmp.EchoMessage> predicate) {
-    try (ScmpResponder responder = Scmp.newResponderBuilder().build()) {
-      responder.setScmpErrorListener(
-          scmpMessage -> errors.add(scmpMessage.getTypeCode().getText()));
-      responder.setOption(ScionSocketOptions.SCION_API_THROW_PARSER_FAILURE, true);
-      responder.setScmpEchoListener(predicate);
-      barrier.countDown();
-      responder.start();
-    } catch (IOException e) {
-      e.printStackTrace();
       throw new RuntimeException(e);
     }
   }
